refactor: move OS window styling into platform/ and type-check keyboard_app
Moved the Win32 window-styling code (WS_EX_NOACTIVATE, always-on-top,
the taskbar-button dance, AppUserModelID, surfacing an existing
instance) into src/platform/windows_window.py, and the pyobjc
activation-policy and NSWindow tuning into src/platform/macos_window.py.
keyboard_app.py now only dispatches to them and is no longer exempt from
mypy in pyproject.toml. Removing the exemption surfaced one real error
(an Optional exc_value passed to Logger.critical's exc_info in the
thread exception hook), which is fixed with a proper null check rather
than a suppression; neither new module needed its own mypy override.
