+++ docs/WHITEPAPER.md
@@ -521,7 +521,7 @@ Every release ships two dependency artefacts alongside the installer, plus a CI-
 
 ### 7.5 macOS (in progress)
 
-A third platform port is scaffolded but not yet shipped. The platform abstraction (`src/platform/macos.py`) implements `MacOSKeySynthesizer` on top of `Quartz.CGEventCreateKeyboardEvent`; the `"win"` modifier maps to ⌘ Command. Window flags (`keyboard_app.py::_apply_macos_window_flags`) set float level, all-Spaces collection behaviour, and `hidesOnDeactivate=NO` so the OSK floats above target apps without joining Mission Control's window list. Config and model directories live under `~/Library/Application Support/alpha-osk/`. The build pipeline at `build/macos/` produces an `Alpha-OSK.app` bundle via PyInstaller `BUNDLE()` with an optional `hdiutil` step for a `.dmg`, but is not yet exercised end-to-end.
+A third platform port is scaffolded but not yet shipped. The platform abstraction (`src/platform/macos.py`) implements `MacOSKeySynthesizer` on top of `Quartz.CGEventCreateKeyboardEvent`; the `"win"` modifier maps to ⌘ Command. Window flags (`src/platform/macos_window.py::apply_window_flags`) set float level, all-Spaces collection behaviour, and `hidesOnDeactivate=NO` so the OSK floats above target apps without joining Mission Control's window list. Config and model directories live under `~/Library/Application Support/alpha-osk/`. The build pipeline at `build/macos/` produces an `Alpha-OSK.app` bundle via PyInstaller `BUNDLE()` with an optional `hdiutil` step for a `.dmg`, but is not yet exercised end-to-end.
 
 The hard part is **input delivery**. Naïve `CGEventPost(kCGHIDEventTap, ev)` posts to whatever app is currently frontmost, and on macOS clicking the OSK window activates the OSK as the foreground app (even with `NSApplicationActivationPolicyAccessory`, `Qt.WindowDoesNotAcceptFocus`, `Qt.Tool`, and `hidesOnDeactivate=NO`, all of which help on other surfaces but do not prevent click-activation of a plain NSWindow in Qt 6). The synthesised keystroke then lands back in Alpha-OSK rather than the editor the user was typing into. The working pattern installs an `NSWorkspaceDidActivateApplicationNotification` observer at synth init, records the pid of every non-self app that activates, and posts each event via `CGEventPostToPid(target_pid, ev)`. Pid-targeted delivery is frontmost-independent, so the event lands in the editor whether or not Alpha-OSK is the foreground app at the instant of dispatch. A cold-start window before any target has been recorded falls back to `CGEventPost`; in practice users tab into a target editor before clicking the OSK, so the fallback is rare.
