+++ docs/architecture/GOTCHAS.md
@@ -41,7 +41,7 @@ A relauncher whose parent PID is already gone still never exits, which is the re
 - **`NgramPredictor._user_total` is an invariant** — every mutation to `user_vocab` (in `learn`, `learn_word`, `_apply_decay`, `clear_user_data`, `load`) must keep it equal to `sum(user_vocab.values())`. `predict()` reads it every keystroke; the consistency tests in `tests/test_ngram_predictor.py::TestUserTotalIncremental` will catch a missed site.
 - **Window height is bound to content — do NOT persist it or assign it imperatively.** `Main.qml` declares `height: outerLayout.implicitHeight + 80`. The `80` budgets for the 48 px title bar's top region plus bottom padding, and it is one of **three coupled numbers** — the `titleBar` `height` (48), `outerLayout.anchors.topMargin` (52 = bar + 4 px gap), and this window-height offset (80) move together; change one without the others and the content either clips the bottom row or floats with a gap. Width (`savedWindowWidth`) and window position (`savedWindowX` / `savedWindowY`, sentinel `-1000000` = never positioned) are restored at startup; height is not. Width and position writes are debounced through `saveGeometryTimer` (300 ms) and all gated on `_geometryRestored` so Qt's construction-time geometry churn isn't persisted before the saved values are restored. `onXChanged` / `onYChanged` restart the timer; the timer's `onTriggered` writes width + x + y together. Position is clamped back on-screen on restore (`Math.max(0, Math.min(saved, Screen.dim - root.dim))`) in case the display layout changed (monitor unplugged, resolution drop). Height stays bound and must never be assigned imperatively. An earlier version also persisted height, which broke the binding the moment `Component.onCompleted` did `root.height = savedWindowHeight` — once the binding was dead, any width change made the keyboard either clip the bottom row or grow empty bands above/below the keys. The user has no vertical resize handle (both edges are `SizeHorCursor`), so width is the only knob; height auto-follows. If you ever need to add height persistence, you also need a re-binding strategy (`Qt.binding(...)`, or an `onHeightChanged` clamp to expected) — don't just assign and walk away.
 - **`KeyButton.qml` drives press visuals off `_visualPressed`, not `mouseArea.pressed`.** With `WS_EX_NOACTIVATE`, Qt occasionally drops the release event when the user drags off the OSK onto another window. Binding visuals straight to `pressed` left the key visually latched. `_visualPressed` is set true on press, cleared on release / cancel / drag-off / a 5 s safety timer — four independent paths back to neutral. If you add a new visual that should follow press state, bind it to `keyRoot._visualPressed`.
-- **Windows taskbar identity must be claimed before the first window.** `keyboard_app.py::_set_windows_app_user_model_id` calls `SetCurrentProcessExplicitAppUserModelID(APP_USER_MODEL_ID)` (`"OKStudio.AlphaOSK"`) once in `main()`, *before* `QApplication` is created. Without it the taskbar button is created at launch with the exe's embedded icon, then reverts to the generic default icon the moment the OSK window appears — Windows re-derives an app identity from the bare process and can't tie the button's icon back to ours. The call must run before any top-level window exists or Windows has already cached the derived identity; it's a no-op off Windows and best-effort on Windows (a failure costs only the taskbar icon, never startup). Don't move it after `QApplication()` / window creation, and keep `APP_USER_MODEL_ID` stable — if the installer shortcuts are ever updated to stamp an AppUserModelID (for pinned-shortcut grouping), they must use this exact string.
+- **Windows taskbar identity must be claimed before the first window.** `src/platform/windows_window.py::set_app_user_model_id` calls `SetCurrentProcessExplicitAppUserModelID(app_id)` (`keyboard_app.py::main()` passes `APP_USER_MODEL_ID`, `"OKStudio.AlphaOSK"`) once in `main()`, *before* `QApplication` is created. Without it the taskbar button is created at launch with the exe's embedded icon, then reverts to the generic default icon the moment the OSK window appears: Windows re-derives an app identity from the bare process and can't tie the button's icon back to ours. The call must run before any top-level window exists or Windows has already cached the derived identity; it's a no-op off Windows and best-effort on Windows (a failure costs only the taskbar icon, never startup). Don't move it after `QApplication()` / window creation, and keep `APP_USER_MODEL_ID` stable: if the installer shortcuts are ever updated to stamp an AppUserModelID (for pinned-shortcut grouping), they must use this exact string.
 - **Single-instance lock holder must outlive `QApplication`.** `keyboard_app.py::_SINGLETON_LOCK` is a module-level `QSharedMemory` reference. `QSharedMemory`'s segment is freed when the holding object is destroyed, so a function-local would release the lock before the app even started. If you refactor `_acquire_singleton_or_surface`, keep that module-level reference alive (or move ownership to the `QApplication` instance).
 - **`KEYBDINPUT.dwExtraInfo` is `ULONG_PTR` (an integer), not a real pointer.** MSDN types it that way; we alias `ULONG_PTR = ctypes.c_size_t` in `src/platform/windows.py`. Set it to `0` for our synthesized input — *never* allocate a Python `c_ulong` and pass `ctypes.pointer()` to it. The kernel doesn't dereference the field, but the Python object would be reaped while the INPUT struct still references its address (a real UB hazard if the field is ever read by another consumer).
 - **Lifetime analytics: persist BOTH session and lifetime.** Any new metric added to `analytics.py` needs (1) a session counter, (2) an `_alltime_*` mirror, (3) load + save in `_load_alltime` / `save`, and (4) a lifetime field in `get_session_stats`. The dashboard's Lifetime / Session toggle expects every visible metric to have both forms; surfacing only session keys regresses the toggle for that tile. `_compute_quality_score` accepts kwargs so the same logic computes session and lifetime scores.
